requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==0.28.1
boto3==1.34.34
python-dotenv==1.0.0
gunicorn==21.2.0
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.2.0
numpy==1.26.3
matplotlib==3.8.2
sentence-transformers==2.3.1
faiss-cpu==1.7.4
scikit-learn==1.4.0
